Reset timer when user drag the scroll view

+++ KIImagePager/KIViewController.m
@@ -65,12 +65,12 @@
 #pragma mark - KIImagePager Delegate
 - (void) imagePager:(KIImagePager *)imagePager didScrollToIndex:(NSUInteger)index
 {
-    NSLog(@"%s %d", __PRETTY_FUNCTION__, index);
+    NSLog(@"%s %lu", __PRETTY_FUNCTION__, (unsigned long)index);
 }
 
 - (void) imagePager:(KIImagePager *)imagePager didSelectImageAtIndex:(NSUInteger)index
 {
-    NSLog(@"%s %d", __PRETTY_FUNCTION__, index);
+    NSLog(@"%s %lu", __PRETTY_FUNCTION__, (unsigned long)index);
 }
 
 @end
